travel/requirements.txt
certifi
chardet
Click
coreapi
coreapi-cli
coreschema
Django
django-filter
djangorestframework
djangorestframework-simplejwt
idna
itypes
Jinja2
Markdown
MarkupSafe
Pillow
psycopg2-binary
pyasn1
pycryptodome
Pygments
PyJWT
python-weixin
pytz
requests
rsa
simplejson
six
sqlparse
uritemplate
urllib3
xmltodict
ckeditor
weixin
django-simpleui